[BUGFIX] The Edit controller confirm action misses autologin
For the confirm action for new accounts the autologin can be
configured. But this misses for the Edit controller.
Also the FeuserEdit->confirm misses from the flexform and
localconf settings.

Added also a default value and extra check for authCode so you
do not get a exception when the authcode is not present in the
link. Can happen when mail client destroys the link.

Change-Id: I1752dd99ff790f1f682c2225f895382c95d0ca8f
Reviewed-on: https://review.typo3.org/26977
Reviewed-by: Sebastian Fischer
Tested-by: Sebastian Fischer