[FEATURE] Allow early persisting during user creation
Currently the user entity is persisted after mail notifications
have been sent, however to use the uid in the notifications,
early persisting would be beneficial.
To avoid that the plain password is persisted, this is cleared
before persisting and restored afterwards. This way, the plain
password still could be used for notifications
(for whatever reason).

Resolves: #60162
Change-Id: I860ddbf340466fa252f99163848c734b868d6d94
Reviewed-on: https://review.typo3.org/31456
Reviewed-by: Sebastian Fischer
Tested-by: Sebastian Fischer