Tweak activation and enable handling in confirm and accept actions
The process is now as following:

A user sets activated on for himself and in case, if an admin
doesn't need to accept the registration, unset disable for himself
too.

An admin can unset disabled for the user and in case, if the user
doesn't need to confirm the registration, set activated on for the
user too.

So now it's possible to:
- User registers -> user confirms
- User registers -> user confirms -> admin accepts
- User registers -> admin accepts
- User registers -> admin accepts -> user confirms
with contextual setting of activated on and unsetting disable

Resolves #76