[TASK] Introduce settings.definitions.yaml
Settings defined by these can be edited in the "Site Management > Settings" module and are used as default value if no TypoScript constant with the same name is filled.

By this its possible to a configure the registration without any TypoScript record and b still override site settings with constants on page level.